+++ views/index.jade
@@ -13,20 +13,17 @@ block content
         th Delete?
       tbody
 
-    // ADD USER
-    h2 Add User
-    #addScore
-      fieldset
-        input#inputName(type='text', placeholder='Name')
-        br
-        input#inputScore(type='text', placeholder='score')
-        br
-        button(type='button', class='btn btn-primary' id='btnAddScore') Add Score
-    // /ADD USER
+  // ADD USER
+  h2 Add Score
+  #addScore
+    fieldset
+      input#inputName(type='text', placeholder='Name')
+      br
+      input#inputScore(type='text', placeholder='score')
+      br
+      button(type='button', class='btn btn-primary' id='btnAddScore') Add Score
 
-    h2 User Info
-    p
-    strong score:
-    | <span id='userInfoscore'></span>
+  p
+  #scoreInfoName
 
-    div(id='game',style='margin:0 auto;')
+  button(type='button', class='btn btn-primary' id='btnPlayGame' hidden=true) Play Game
