+++ views/layout.jade
@@ -17,6 +17,10 @@ html
     script(type="text/javascript", src='https://rawgithub.com/craftyjs/Crafty/release/dist/crafty-min.js')
 
     script(src='javascripts/main.js')
+    script(src='javascripts/DataLogic/mongo.js')
+    script(src='javascripts/Game/components.js')
+    script(src='javascripts/Game/game.js')
+    script(src='javascripts/Game/scenes.js')
 
   body
     block content
